requirements.txt
Flask
gunicorn
pandas
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
build